grub: automatically substitute uuid in kernel cmdline
The uuid of the $ROOT device is the uuid of the partition mounted
by bootchain or propagator. Therefore, we can substitute it, and
not specify it in advance. This will allow you to transfer, as is,
the contents of the iso to the flash drive.
Also, for propagator, we replace the method with disk. This should
not cause any problems.